+++ Src/SCPCB/Src/Map/Doors.cpp
@@ -23,7 +23,7 @@ Door::Door() {
     list.push_back(this);
 }
 Door::~Door() {
-    for (int i = 0; i < list.size(); i++) {
+    for (int i = 0; i < (int)list.size(); i++) {
         if (list[i] == this) {
             list.erase(list.begin() + i);
             break;
