+++ Src/SCPCB/Src/Map/Objects.cpp
@@ -22,7 +22,7 @@ Lever::Lever() {
     list.push_back(this);
 }
 Lever::~Lever() {
-    for (int i = 0; i < list.size(); i++) {
+    for (int i = 0; i < (int)list.size(); i++) {
         if (list[i] == this) {
             list.erase(list.begin() + i);
             break;
@@ -41,7 +41,7 @@ ElevatorObj::ElevatorObj() {
     list.push_back(this);
 }
 ElevatorObj::~ElevatorObj() {
-    for (int i = 0; i < list.size(); i++) {
+    for (int i = 0; i < (int)list.size(); i++) {
         if (list[i] == this) {
             list.erase(list.begin() + i);
             break;
