+++ Src/SCPCB/Src/NPCs/NPCtype106.cpp
@@ -22,7 +22,7 @@ Data106::Data106() {
     list.push_back(this);
 }
 Data106::~Data106() {
-    for (int i = 0; i < list.size(); i++) {
+    for (int i = 0; i < (int)list.size(); i++) {
         if (list[i] == this) {
             list.erase(list.begin() + i);
             break;
